// frontend/xacria-presenze/src/generated-client/model/calendarWorkingDayEntry.ts
/**
 * API Presenze-Pro
 * API used by the Presenze-Pro thesis project’s client
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Rappresenta una entry riguardante una task giornaliera nel calendario
 */
export interface CalendarWorkingDayEntry { 
    project?: string;
    hourFrom?: string;
    hourTo?: string;
    dateFrom?: Date;
}